+++ src/main/java/com/alibou/banking/account/Account.java
@@ -1,12 +1,11 @@
 package com.alibou.banking.account;
 
 
+import com.alibou.banking.common.AbstractEntity;
 import com.alibou.banking.user.User;
 
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.JoinColumn;
@@ -15,7 +14,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
+import lombok.experimental.SuperBuilder;
 
 
 @Entity
@@ -23,11 +22,10 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "ACCOUNT")
-public class Account {
-    @Id
-    @GeneratedValue
-    private Long id;
+@SuperBuilder
+@Table(name = "ACCOUNTS")
+public class Account extends AbstractEntity {
+
     private String iban;
     @OneToOne
     @JoinColumn(name = "user_id")
