+++ src/main/java/com/alibou/banking/address/Address.java
@@ -1,15 +1,12 @@
 package com.alibou.banking.address;
 
 
-
-
+import com.alibou.banking.common.AbstractEntity;
 import com.alibou.banking.user.User;
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -20,11 +17,9 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "ADDRESS")
-public class Address {
-    @Id
-    @GeneratedValue
-    private Long id;
+@Table(name = "ADDRESSES")
+public class Address extends AbstractEntity {
+
     private String street;
     private String city;
     private String state;
