+++ src/main/java/com/alibou/banking/fraud/Fraud.java
@@ -1,12 +1,11 @@
 package com.alibou.banking.fraud;
 
+import com.alibou.banking.common.AbstractEntity;
 import com.alibou.banking.transaction.Transaction;
 
 
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.EnumType;
@@ -15,6 +14,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
 import java.time.LocalDateTime;
 
 
@@ -23,11 +24,10 @@ import java.time.LocalDateTime;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "FRAUD")
-public class Fraud {
-    @Id
-    @GeneratedValue
-    private Long id;
+@SuperBuilder
+@Table(name = "FRAUDS")
+public class Fraud extends AbstractEntity {
+
     @Enumerated(EnumType.STRING)
     private FraudType type;
     @Enumerated(EnumType.STRING)
