+++ src/main/java/com/alibou/banking/transaction/Transaction.java
@@ -1,5 +1,6 @@
 package com.alibou.banking.transaction;
 
+import com.alibou.banking.common.AbstractEntity;
 import com.alibou.banking.fraud.Fraud;
 
 import com.alibou.banking.user.User;
@@ -9,8 +10,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Table;
-import jakarta.persistence.Id;
-import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.JoinColumn;
@@ -19,6 +18,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Setter;
+import lombok.experimental.SuperBuilder;
+
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
@@ -27,11 +28,10 @@ import java.time.LocalDateTime;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Table(name = "TRANSACTION")
-public class Transaction {
-    @Id
-    @GeneratedValue
-    private Long id;
+@SuperBuilder
+@Table(name = "TRANSACTIONS")
+public class Transaction extends AbstractEntity {
+
     private String description;
     private BigDecimal amount;
     private LocalDateTime date;
